+++ R/downloadablePlot.R
@@ -131,12 +131,15 @@ downloadablePlotUI <- function(id,
 
     }
 
-    btn_item <- shiny::span(id    = ns("dplotButtonDiv"),
-                            class = "periscope-downloadable-plot-button",
-                            style = styleval,
-                            periscope2::downloadFileButton(ns("dplotButtonID"),
-                                                           downloadtypes,
-                                                           download_hovertext))
+    btn_item <- shiny::conditionalPanel(
+        condition = "output.displayButton",
+        ns        = ns,
+        shiny::span(id    = ns("dplotButtonDiv"),
+                    class = "periscope-downloadable-plot-button",
+                    style = styleval,
+                    periscope2::downloadFileButton(ns("dplotButtonID"),
+                                                   downloadtypes,
+                                                   download_hovertext)))
 
     plot_item <- shiny::plotOutput(outputId = ns("dplotOutputID"),
                                    width    = width,
@@ -226,10 +229,10 @@ downloadablePlotUI <- function(id,
 #'
 #' @export
 downloadablePlot <- function(id,
-                             logger,
-                             filenameroot,
+                             logger       = NULL,
+                             filenameroot = "download",
                              aspectratio  = 1,
-                             downloadfxns = list(),
+                             downloadfxns = NULL,
                              visibleplot) {
     shiny::moduleServer(
         id,
@@ -250,14 +253,17 @@ downloadablePlot <- function(id,
             })
 
             shiny::observe({
-                if (!is.null(downloadfxns) && (length(downloadfxns) > 0)) {
+                if (length(downloadfxns) > 0) {
                     dpInfo$downloadfxns <- lapply(downloadfxns, do.call, list())
-                    rowct <- lapply(dpInfo$downloadfxns, is.null)
+                    rowct               <- lapply(dpInfo$downloadfxns, is.null)
                     session$sendCustomMessage(
                         "downloadbutton_toggle",
                         message = list(btn  = session$ns("dplotButtonDiv"),
                                        rows = sum(unlist(rowct) == FALSE)) )
                 }
+
+                output$displayButton <- reactive(length(downloadfxns) > 0)
+                outputOptions(output, "displayButton", suspendWhenHidden = FALSE)
             })
         })
 }